data-analytics/knowledge_graph_demo/manufacturing_assistant_agent/requirements.txt
google-cloud-aiplatform[agent_engines,adk]==1.133.0
google-adk==1.22.1
google-genai==1.59.0
google-cloud-bigquery
google-auth
python-dotenv